Add Rust winit crate as alternative to GLFW (for future Android support) (#545)
* WIP

* Let there be a window!

* Initial cleanup

* Renamed sp-rust crate to sp-sys

* sp-sys crate

* Sync Tecs

* Formatting

* Removed clang dependency for rust build

* Windows build fixes

* WIP: Initial demo with Android build support

* WIP: c++ bridge and android demo

* Spellcheck

* Android build support

* Updated vscode extension for rust crates

* cxx bridge build fix

* default desktop gen for rust demo

* Add strayphotons.h header

* WIP: Rust bindings to libsp

* Rust build warning cleanup

* Cmake linking updates for sp.lib

* Fix libsp linking issues

* WIP: dynamic linking

* WIP: rust debugging

* Fix cargo build on Windows

* Fix command line arguments crashing, and catch and print exceptions

* Make new vulkan-desktop target

* Restore GLFW input and hook up new Winit CMake modules

* Add back in missing rust/ files. Improve Winit input handler

* Fix and re-enable precompiled headers

* Minor fixes

* Removed stdlib from clang args

* Update Tecs, move strayphotons.h header, improve link search path

* CI fix and code cleanup

* Fix windows build linking

* Implement winit keyboard and mouse input
Fix shutdown ordering crash
Clean up winit/gfx module

* Cleanup and enable debug utils messages in rust

* Add monitor mode reading to winit impl, fix crt-static on windows

* Fix Winit shutdown ordering
Pipe through exit codes from C++ to Rust
Fix Winit mouse grabbing
Fix cli arg passing from Winit to C++
Implement set_window_title for Winit
Win32 build fixes

* Fix sp-test execution

* Fix winSchedulerHandle breaking on linux

* Read win32 handle for ImGui, enable audio in cargo build for windows

* Implement fullscreen interface mode

* WIP: Fixes

* Build sp-rs-wasm and sp-rs-window configurations separately

* Fix build dependencies

* Fix windows build

* Update vulkano and winit to non-beta versions

* Strip out android code for now

* Review updates

* More review changes

* Add missing license comment

---------

+++ src/core/ecs/SignalExpression.cc
@@ -924,7 +924,7 @@ namespace ecs {
     double SignalExpression::Evaluate(const DynamicLock<ReadSignalsLock> &lock, size_t depth) const {
         // ZoneScoped;
         // ZoneStr(expr);
-        if (rootIndex < 0 || rootIndex >= nodes.size()) return 0.0;
+        if (rootIndex < 0 || (size_t)rootIndex >= nodes.size()) return 0.0;
         Storage cache;
         auto &rootNode = nodes[rootIndex];
         Context ctx(lock, *this, cache, 0.0);
@@ -934,7 +934,7 @@ namespace ecs {
     double SignalExpression::EvaluateEvent(const DynamicLock<ReadSignalsLock> &lock, const EventData &input) const {
         // ZoneScoped;
         // ZoneStr(expr);
-        if (rootIndex < 0 || rootIndex >= nodes.size()) return 0.0;
+        if (rootIndex < 0 || (size_t)rootIndex >= nodes.size()) return 0.0;
         Storage cache;
         auto &rootNode = nodes[rootIndex];
         Context ctx(lock, *this, cache, input);
